requirements.txt
fastapi
uvicorn
langchain-groq
langchain
python-dotenv
requests
pydub
elevenlabs
freesound-python
huggingface_hub
Pillow
freesound-api